requirements.txt
setuptools>=0.8
django==1.5.4
django-nose==1.2
mock==1.0.1
factory-boy==2.1.2
pytz==2013d
